split reducers and actions into separate files
