+++ react/src/model/auth/index.js
@@ -1,0 +1,7 @@
+import reducer from './reducer'
+import { loadMnemonic } from './actions'
+
+export {
+  reducer as default,
+  loadMnemonic
+}
