+++ react/src/model/notes/index.js
@@ -1,0 +1,8 @@
+import reducer from './reducer'
+import { fetchNotes, createNote } from './actions'
+
+export {
+  reducer as default,
+  fetchNotes,
+  createNote
+}
