+++ react/src/model/users/index.js
@@ -1,0 +1,8 @@
+import reducer from './reducer'
+import { fetchUsers, createUser } from './actions'
+
+export {
+  reducer as default,
+  fetchUsers,
+  createUser
+}
